Set up backups
aw yea tarsnap

pretty ideal: it's really cheap for the amount of storage used, sqlite3 seems good for dedup (and can switch to ascii dump if necessary), and I can load the account with a year+ of credit for minimal maintenance burden.

I think the primary downide is that the account suspended period is just seven days. It will probably be a good idea to track the credit to predict when it will reach zero, to give a bigger re-deposit window of time.

It is hosted in the US on AWS, unfortunately. However, since the client encryption software seems trustworthy, it feels like the right tradeoff in this case. They have an EU region host available for GDPR compliance, so, maybe I can get on that later.