remove httpd fallback in relay conf
since the request gets matched and forwarded in the `protocol` anyway, it's not needed.

having it there seemed to cause occasional problems, as relayd would sometimes forward it requests meant for the `commit--blog` app, to which it would always reply with `404`.

with it gone, it seems like relayd tries harder to make it work with the actual app backend. if it ever can't reach the backend, we'll hopefully see an appropriate `5xx` error instead of the incorrect `4xx` ones i've been seeing.