install system py3-cryptography
latest pypi version fails because pip doesn't know how to build with rust. older pypi version fails because openssl is not here.

but it's on openbsd which tbh sounds better. pip just needs some convincing to chill and believe that things will be fine. handily, all the deps are explictly frozen with pip in the `requirements.txt` generated over at commit--blog, so adding `--no-deps` here should be alright.